command: improve the randomkey command
It is not random in fact, it is difficult to return a key in a really
random way. We use a two-step random strategy here to return a key, but
it's distribution is not uniform. The main propose of this command is to
return a key to inspect the database, so the random guarantee is not
always necessary.